Add files via upload
Electronic health records are one of the essential components of health organizations. In recent years, there have been increased concerns about privacy and reputation regarding the storage and use of patient information. In this regard, the information provided as a part of medical and health insurance, for instance, can be viewed as proof of social insurance and governance. Several problems in the past few decades regarding medical information management have threatened patient information privacy. In intelligent healthcare applications, the privacy of patients' data is one of the main concerns. As a result, Blockchain is a severe necessity as it can enhance transparency and security in medical applications. Accordingly, this paper uses the design science method to propose a secure Blockchain framework for healthcare records management systems. The proposed framework comprises five components: a Blockchain network, smart contracts, privacy key management, data encryption, and integration with healthcare information technology. In the proposed framework, healthcare organizations can manage healthcare information securely and privately. Additionally, a secure storage system for electronic records is proposed to meet these organizations' needs. It provides security and privacy for healthcare organizations, especially when managing healthcare information, and also proposes a secure storage system for electronic records to meet the needs of the organizations.